refac: 기존 GameService -> GameHistoryService, 새로운 GameService 생성
