+++ src/game/game.controller.ts
@@ -10,7 +10,7 @@ import {
   UseGuards,
 } from '@nestjs/common';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
-import { GameService } from './game.service';
+import { GameHistoryService } from './game-history.service';
 import { CreateGameDto } from './dto/create-game.dto';
 import { GameRecordService } from 'src/gameRecord/game-record.service';
 import { GameType } from './const/game.type';
@@ -19,7 +19,7 @@ import { GameType } from './const/game.type';
 @UseGuards(JwtAuthGuard)
 export class GameController {
   constructor(
-    private readonly gameService: GameService,
+    private readonly gameHistoryService: GameHistoryService,
     private readonly gameRecordService: GameRecordService,
   ) {}
   @Get('rank')
@@ -27,7 +27,7 @@ export class GameController {
     @Query('id') id: number,
     @Query('offset') offset: number,
   ) {
-    return this.gameService.getAllGamesById(id, GameType.RANK, offset);
+    return this.gameHistoryService.getAllGamesById(id, GameType.RANK, offset);
   }
 
   @Get('general')
@@ -35,13 +35,13 @@ export class GameController {
     @Query('id') id: number,
     @Query('offset') offset: number,
   ) {
-    return this.gameService.getAllGamesById(id, GameType.NORMAL, offset);
+    return this.gameHistoryService.getAllGamesById(id, GameType.NORMAL, offset);
   }
 
   @Post('game')
   @HttpCode(HttpStatus.CREATED)
   async createGame(@Body() createGameDto: CreateGameDto) {
-    await this.gameService.createGame(createGameDto);
+    await this.gameHistoryService.createGame(createGameDto);
     await this.gameRecordService.updateGameRecord(
       createGameDto.winnerId,
       createGameDto.loserId,
@@ -51,11 +51,11 @@ export class GameController {
   //삭제 예정
   @Get('all')
   async getGameAll() {
-    return this.gameService.findAll();
+    return this.gameHistoryService.findAll();
   }
   //삭제 예정
   @Post(':id')
   async getGameById(@Param('id') id: number) {
-    return this.gameService.findByUserId(id);
+    return this.gameHistoryService.findByUserId(id);
   }
 }
